requirements.txt
fastapi
uvicorn
pandas
openpyxl
PyPDF2
gdown
requests
python-dotenv
langchain-groq
langgraph
pydantic
python-multipart